LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/no-spacing-expected.txt


FAIL Spacing inside an <mfrac> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 227
PASS Spacing around an <mfrac> child of an <math>.
FAIL Spacing inside an <mfrac> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 227
PASS Spacing around an <mfrac> child of an <mrow>.
FAIL Spacing inside an <mfrac> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 227
PASS Spacing around an <mfrac> child of an <mtd>.
FAIL Spacing inside an <msub> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msub> child of an <math>.
FAIL Spacing inside an <msub> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msub> child of an <mrow>.
FAIL Spacing inside an <msub> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msub> child of an <mtd>.
FAIL Spacing inside an <msup> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msup> child of an <math>.
FAIL Spacing inside an <msup> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msup> child of an <mrow>.
FAIL Spacing inside an <msup> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msup> child of an <mtd>.
FAIL Spacing inside an <msubsup> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msubsup> child of an <math>.
FAIL Spacing inside an <msubsup> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msubsup> child of an <mrow>.
FAIL Spacing inside an <msubsup> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 451.03125
PASS Spacing around an <msubsup> child of an <mtd>.
FAIL Spacing inside an <mmultiscripts> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 677.0625
PASS Spacing around an <mmultiscripts> child of an <math>.
FAIL Spacing inside an <mmultiscripts> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 677.0625
PASS Spacing around an <mmultiscripts> child of an <mrow>.
FAIL Spacing inside an <mmultiscripts> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 677.0625
PASS Spacing around an <mmultiscripts> child of an <mtd>.
FAIL Spacing inside an <munder> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munder> child of an <math>.
FAIL Spacing inside an <munder> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munder> child of an <mrow>.
FAIL Spacing inside an <munder> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munder> child of an <mtd>.
FAIL Spacing inside an <mover> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <mover> child of an <math>.
FAIL Spacing inside an <mover> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <mover> child of an <mrow>.
FAIL Spacing inside an <mover> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <mover> child of an <mtd>.
FAIL Spacing inside an <munderover> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munderover> child of an <math>.
FAIL Spacing inside an <munderover> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munderover> child of an <mrow>.
FAIL Spacing inside an <munderover> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 225
PASS Spacing around an <munderover> child of an <mtd>.
FAIL Spacing inside an <mroot> child of <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 468.0625
FAIL Spacing around an <mroot> child of an <math>. assert_less_than_equal: expected a number less than or equal to 100 but got 468.0625
FAIL Spacing inside an <mroot> child of <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 468.0625
FAIL Spacing around an <mroot> child of an <mrow>. assert_less_than_equal: expected a number less than or equal to 100 but got 468.0625
FAIL Spacing inside an <mroot> child of <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 468.0625
FAIL Spacing around an <mroot> child of an <mtd>. assert_less_than_equal: expected a number less than or equal to 100 but got 488.0625
X
X

X
X

X
X


X
X
X
X
X
X


X
X
X
X
X
X


X
X
X
X
X
X
X
X
X


X
X
X
X
X
X
X
X
X
X
X
X
X
X
X


X
X

X
X

X
X


X
X

X
X

X
X


X
X
X

X
X
X

X
X
X


X
X
X
X
X
X
